+++ Com.Adobe.Xmp/Com/adobe/xmp/impl/Utils.cs
@@ -375,7 +375,7 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Returns true if the char is an ASCII control char.</returns>
         internal static bool IsControlChar(char c)
         {
-            return (c <= unchecked((int)(0x1F)) || c == unchecked((int)(0x7F))) && c != unchecked((int)(0x09)) && c != unchecked((int)(0x0A)) && c != unchecked((int)(0x0D));
+            return (c <= unchecked(0x1F) || c == unchecked(0x7F)) && c != unchecked(0x09) && c != unchecked(0x0A) && c != unchecked(0x0D);
         }
 
         /// <summary>Serializes the node value in XML encoding.</summary>
@@ -487,10 +487,9 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Returns true if the character is a valid first char of an XML name.</returns>
         private static bool IsNameStartChar(char ch)
         {
-            return (ch <= unchecked((int)(0xFF)) && _xmlNameStartChars[ch]) || (ch >= unchecked((int)(0x100)) && ch <= unchecked((int)(0x2FF))) || (ch >= unchecked((int)(0x370)) && ch <= unchecked((int)(0x37D))) || (ch >= unchecked((int)(0x37F)) && ch <=
-                 unchecked((int)(0x1FFF))) || (ch >= unchecked((int)(0x200C)) && ch <= unchecked((int)(0x200D))) || (ch >= unchecked((int)(0x2070)) && ch <= unchecked((int)(0x218F))) || (ch >= unchecked((int)(0x2C00)) && ch <= unchecked((int)(0x2FEF))) ||
-                (ch >= unchecked((int)(0x3001)) && ch <= unchecked((int)(0xD7FF))) || (ch >= unchecked((int)(0xF900)) && ch <= unchecked((int)(0xFDCF))) || (ch >= unchecked((int)(0xFDF0)) && ch <= unchecked((int)(0xFFFD))) || (ch >= unchecked((int)(0x10000
-                )) && ch <= unchecked((int)(0xEFFFF)));
+            return (ch <= unchecked(0xFF) && _xmlNameStartChars[ch]) || (ch >= unchecked(0x100) && ch <= unchecked(0x2FF)) || (ch >= unchecked(0x370) && ch <= unchecked(0x37D)) || (ch >= unchecked(0x37F) && ch <=
+                 unchecked(0x1FFF)) || (ch >= unchecked(0x200C) && ch <= unchecked(0x200D)) || (ch >= unchecked(0x2070) && ch <= unchecked(0x218F)) || (ch >= unchecked(0x2C00) && ch <= unchecked(0x2FEF)) ||
+                (ch >= unchecked(0x3001) && ch <= unchecked(0xD7FF)) || (ch >= unchecked(0xF900) && ch <= unchecked(0xFDCF)) || (ch >= unchecked(0xFDF0) && ch <= unchecked(0xFFFD)) || (ch >= unchecked(0x10000) && ch <= unchecked(0xEFFFF));
         }
 
         /// <summary>
@@ -502,7 +501,7 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Returns true if the character is a valid char of an XML name.</returns>
         private static bool IsNameChar(char ch)
         {
-            return (ch <= unchecked((int)(0xFF)) && _xmlNameChars[ch]) || IsNameStartChar(ch) || (ch >= unchecked((int)(0x300)) && ch <= unchecked((int)(0x36F))) || (ch >= unchecked((int)(0x203F)) && ch <= unchecked((int)(0x2040)));
+            return (ch <= unchecked(0xFF) && _xmlNameChars[ch]) || IsNameStartChar(ch) || (ch >= unchecked(0x300) && ch <= unchecked(0x36F)) || (ch >= unchecked(0x203F) && ch <= unchecked(0x2040));
         }
 
         /// <summary>
@@ -512,13 +511,13 @@ namespace Com.Adobe.Xmp.Impl
         /// </summary>
         private static void InitCharTables()
         {
-            _xmlNameChars = new bool[unchecked((int)(0x0100))];
-            _xmlNameStartChars = new bool[unchecked((int)(0x0100))];
+            _xmlNameChars = new bool[unchecked(0x0100)];
+            _xmlNameStartChars = new bool[unchecked(0x0100)];
             for (char ch = (char)0; ch < _xmlNameChars.Length; ch++)
             {
-                _xmlNameStartChars[ch] = ch == ':' || ('A' <= ch && ch <= 'Z') || ch == '_' || ('a' <= ch && ch <= 'z') || (unchecked((int)(0xC0)) <= ch && ch <= unchecked((int)(0xD6))) || (unchecked((int)(0xD8)) <= ch && ch <= unchecked((int)(0xF6))) || (unchecked(
-                    (int)(0xF8)) <= ch && ch <= unchecked((int)(0xFF)));
-                _xmlNameChars[ch] = _xmlNameStartChars[ch] || ch == '-' || ch == '.' || ('0' <= ch && ch <= '9') || ch == unchecked((int)(0xB7));
+                _xmlNameStartChars[ch] = ch == ':' || ('A' <= ch && ch <= 'Z') || ch == '_' || ('a' <= ch && ch <= 'z') || (unchecked(0xC0) <= ch && ch <= unchecked(0xD6)) || (unchecked(0xD8) <= ch && ch <= unchecked(0xF6)) || (unchecked(
+                    0xF8) <= ch && ch <= unchecked(0xFF));
+                _xmlNameChars[ch] = _xmlNameStartChars[ch] || ch == '-' || ch == '.' || ('0' <= ch && ch <= '9') || ch == unchecked(0xB7);
             }
         }
     }
