+++ Com.Adobe.Xmp/Com/adobe/xmp/impl/XmpUtils.cs
@@ -797,7 +797,7 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Return the character kind.</returns>
         private static int ClassifyCharacter(char ch)
         {
-            if (Spaces.IndexOf(ch) >= 0 || (unchecked((int)(0x2000)) <= ch && ch <= unchecked((int)(0x200B))))
+            if (Spaces.IndexOf(ch) >= 0 || (unchecked(0x2000) <= ch && ch <= unchecked(0x200B)))
             {
                 return UckSpace;
             }
@@ -809,11 +809,11 @@ namespace Com.Adobe.Xmp.Impl
             {
                 return UckSemicolon;
             }
-            if (Quotes.IndexOf(ch) >= 0 || (unchecked((int)(0x3008)) <= ch && ch <= unchecked((int)(0x300F))) || (unchecked((int)(0x2018)) <= ch && ch <= unchecked((int)(0x201F))))
+            if (Quotes.IndexOf(ch) >= 0 || (unchecked(0x3008) <= ch && ch <= unchecked(0x300F)) || (unchecked(0x2018) <= ch && ch <= unchecked(0x201F)))
             {
                 return UckQuote;
             }
-            if (ch < unchecked((int)(0x0020)) || Controls.IndexOf(ch) >= 0)
+            if (ch < unchecked(0x0020) || Controls.IndexOf(ch) >= 0)
             {
                 return UckControl;
             }
@@ -1019,7 +1019,7 @@ namespace Com.Adobe.Xmp.Impl
         /// <returns>Returns true if the character is a closing quote.</returns>
         private static bool IsClosingingQuote(char ch, char openQuote, char closeQuote)
         {
-            return ch == closeQuote || (openQuote == unchecked((int)(0x301D)) && ch == unchecked((int)(0x301E)) || ch == unchecked((int)(0x301F)));
+            return ch == closeQuote || (openQuote == unchecked(0x301D) && ch == unchecked(0x301E) || ch == unchecked(0x301F));
         }
 
         /// <summary>
